Retrieving all agents in MainApp isntead of app. Using connect in App messes with react-router
